Move model modules to top-level.
